requirements.txt
Pillow>=10.0.0
playwright>=1.40.0
Flask>=3.0.0
requests>=2.31.0
moviepy>=2.0.0
edge-tts>=7.0.0
